Add permissions command to verify
