+++ docs/moactl_verify_permissions.md
@@ -1,0 +1,38 @@
+## moactl verify permissions
+
+Verify AWS permissions are ok for cluster install
+
+### Synopsis
+
+Verify AWS permissions are ok for cluster install
+
+```
+moactl verify permissions [flags]
+```
+
+### Examples
+
+```
+  # Verify resources needed to create a cluster are configured as expected
+
+  # Verify AWS permissions are configured correctly
+  moactl verify permissions
+```
+
+### Options
+
+```
+  -h, --help   help for permissions
+```
+
+### Options inherited from parent commands
+
+```
+      --debug     Enable debug mode.
+  -v, --v Level   log level for V logs
+```
+
+### SEE ALSO
+
+* [moactl verify](moactl_verify.md)	 - Verify resources are configured correctly for cluster install
+
